+++ src/post_modules/Shuffler/a_source_system_memory.c
@@ -31,7 +31,7 @@ struct sm_source
   int offset, len;
 };
 
-static struct data get_data( struct source *_s, int len )
+static struct data get_data( struct source *_s, off_t len )
 {
   struct sm_source *s = (struct sm_source *)_s;
   struct data res;
